+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Upcoming
+* Added an option to make mobs find a better path to the target
 * Entities with higher follow range than the override are no longer overridden (e.g. Ghasts)
 * Creepers now start exploding when nearer to the target (explosion radius ~~x1.5~~ -> 1.35 blocks)
 * Cleaned up Creeper Module and AI, should perform slightly better